improved handling of path argument
previously just accessed files correctly based on the provided path,
but now we actively chdir() into the specified path on startup. This
means that things like walking directories will yield consistent
filepaths regardless of whether the program was directly invoked
in a directory, or whether it was asked to run there using --path.