factored common column writing into local fn
During CSV export the code that writes various properties of
(a target | the overall results) to various columns was duplicated
(specifying each column multiple times). This has been factored
out, specifically in view of how new columns are coming along. It
would be too easy to forget a new column being added in one of
the two places.