bugfix: ensure pull done after fetching
when updating repositories, if already on the correct branch after
a fetch the previous code took no further action. this left the
local checkout with an up to date internal blob database but
a working directory containing out-of-date code!

now, after a fetch - if already on the right branch, etc. - a pull
is done to ensure the branch is up to date.